+++ src/Exception/NetworkTimeoutException.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace GuzzleHttp\Exception;
 
 /**
- * Exception thrown when a transfer times out before a response is received.
+ * Exception thrown when a transfer times out after connection is established
+ * but before a response is received.
  */
-class NetworkTimeoutException extends NetworkException
+class NetworkTimeoutException extends NetworkException implements TimeoutException
 {
 }