+++ src/Exception/TimeoutException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GuzzleHttp\Exception;
+
+interface TimeoutException extends GuzzleException
+{
+}