+++ tests/Exception/ConnectTimeoutExceptionTest.php
@@ -15,20 +15,21 @@ use Psr\Http\Client\NetworkExceptionInterface;
 use Psr\Http\Client\RequestExceptionInterface;
 
 /**
- * @covers \GuzzleHttp\Exception\ConnectException
+ * @covers \GuzzleHttp\Exception\ConnectTimeoutException
  */
-class ConnectExceptionTest extends TestCase
+class ConnectTimeoutExceptionTest extends TestCase
 {
     public function testHasRequest(): void
     {
         $req = new Request('GET', '/');
         $prev = new \Exception();
-        $e = new ConnectException('foo', $req, $prev, ['foo' => 'bar']);
+        $e = new ConnectTimeoutException('foo', $req, $prev, ['foo' => 'bar']);
+
+        self::assertInstanceOf(ConnectException::class, $e);
         self::assertInstanceOf(NetworkException::class, $e);
         self::assertInstanceOf(NetworkExceptionInterface::class, $e);
-        self::assertNotInstanceOf(ConnectTimeoutException::class, $e);
+        self::assertInstanceOf(TimeoutException::class, $e);
         self::assertNotInstanceOf(NetworkTimeoutException::class, $e);
-        self::assertNotInstanceOf(TimeoutException::class, $e);
         self::assertNotInstanceOf(RequestExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());
         self::assertSame('foo', $e->getMessage());