+++ tests/Exception/NetworkTimeoutExceptionTest.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace GuzzleHttp\Tests\Exception;
 
 use GuzzleHttp\Exception\ConnectException;
+use GuzzleHttp\Exception\ConnectTimeoutException;
 use GuzzleHttp\Exception\NetworkException;
 use GuzzleHttp\Exception\NetworkTimeoutException;
+use GuzzleHttp\Exception\TimeoutException;
 use GuzzleHttp\Psr7\Request;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Client\NetworkExceptionInterface;
@@ -25,7 +27,9 @@ class NetworkTimeoutExceptionTest extends TestCase
 
         self::assertInstanceOf(NetworkException::class, $e);
         self::assertInstanceOf(NetworkExceptionInterface::class, $e);
+        self::assertInstanceOf(TimeoutException::class, $e);
         self::assertNotInstanceOf(ConnectException::class, $e);
+        self::assertNotInstanceOf(ConnectTimeoutException::class, $e);
         self::assertNotInstanceOf(RequestExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());
         self::assertSame('foo', $e->getMessage());